feat: 회원가입 API 호출 로직을 별도의 함수로 분리하여 코드 정리
